chore: clean up types
