+++ src/notebooks/pipes/Visualization/view.tsx
@@ -1,7 +1,12 @@
 // Libraries
-import React, {FC, useContext} from 'react'
+import React, {FC, useContext, useCallback, useState} from 'react'
 
 // Components
+import {
+  SquareButton,
+  IconFont,
+  ComponentColor
+} from '@influxdata/clockface'
 import EmptyQueryView, {ErrorFormat} from 'src/shared/components/EmptyQueryView'
 import DashboardList from './DashboardList'
 import ViewSwitcher from 'src/shared/components/ViewSwitcher'
@@ -15,9 +20,8 @@ import ExportVisualizationButton from 'src/notebooks/pipes/Visualization/ExportV
 import {event} from 'src/cloud/utils/reporting'
 
 // Types
-import {PipeProp, PipeData} from 'src/notebooks'
+import {PipeProp, PipeData} from 'src/types/notebooks'
 import {ViewType} from 'src/types'
-import {IconFont} from '@influxdata/clockface'
 import {FromFluxResult} from '@influxdata/giraffe'
 
 // NOTE we dont want any pipe component to be directly dependent
@@ -96,6 +100,10 @@ export {updateVisualizationType}
 const Visualization: FC<PipeProp> = ({Context}) => {
   const {timeZone} = useContext(AppSettingContext)
   const {data, update, loading, results} = useContext(PipeContext)
+  const [optionsVisibility, setOptionsVisibility] = useState(false)
+  const toggleOptions = useCallback(() => {
+      setOptionsVisibility(!optionsVisibility)
+  }, [optionsVisibility, setOptionsVisibility])
 
   const updateType = (type: ViewType) => {
     event('Notebook Visualization Type Changed', {
@@ -110,6 +118,13 @@ const Visualization: FC<PipeProp> = ({Context}) => {
       <ViewTypeDropdown
         viewType={data.properties.type}
         onUpdateType={updateType as any}
+      />
+      <SquareButton
+        icon={IconFont.CogThick}
+        onClick={toggleOptions}
+        color={optionsVisibility ? ComponentColor.Primary : ComponentColor.Default}
+        titleText="Configure Visualization"
+        className="flows-config-visualization-button"
       />
       <ExportVisualizationButton disabled={!results.source}>
         {onHidePopover => (
